refactor(routing): Remove compose routing, use a kobweb layout in pages.
